requirements.txt
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
boto3==1.12.9
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dash==1.8.0
dash-core-components==1.7.0
dash-html-components==1.0.2
dash-renderer==1.2.3
dash-table==4.6.0
decorator==4.4.1
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
eventlet==0.25.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
futures==3.1.1
ghostscript==0.6
greenlet==0.4.15
idna==2.9
importlib-metadata==1.1.0
ipdb==0.12.3
ipykernel==5.1.3
ipython==7.10.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.6.1
lxml==4.5.0
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
more-itertools==8.0.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.1
numpy==1.18.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
path==13.1.0
path.py==12.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.5.2
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyrsistent==0.15.6
pytesseract==0.3.2
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
retrying==1.3.3
s3transfer==0.3.3
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.13.0
soupsieve==2.0
SQLAlchemy==1.3.12
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==0.6.0
